Add automated test-generation scripts, generate p.cnt test
pulp_test_gen is a Python package written to simplify the process of
creating future tests for (arithmetic) instructions.
It abstracts the assembler macros and provides different classes based
on the type of operands used in the instruction.
In order to use it, one needs to create an inherited class from the
correct type/class, set operand-limits and implement the operation.
Then this new instruction-specific class can be used to manually add
functional tests or generate a number of random tests.
At the end the testflow can be written to an assembler file.

gen_test_p_cnt.py serves as a rather simple example on how to
use the package.