Add tests for hwloop subset of Xpulpimg
The expected effects of endi, starti, count and counti are tested
directly (they should write to CSR, CSR is read-back to check).

Since setup and setupi immediately start a loop, this is not really
possible. Therefore they are tested by using them in a small loop.
This is not at all exhaustive.
In the future more complicated setups/loops should be added.