Add test for bitrev subset (not in Makefile)
This test did not yet run against the Spike implementation because
xpulpbitrev is not an extension in gcc and xgap9 throws an error.

The generated tests are partially verified by hand, additionally the
first three are examples from the docs and they are correct.
I am fairly certain the test is fine.